requirements.txt
flair==0.13.1
scipy<=1.10.1
bs4==0.0.2
PyYAML==6.0.1
requests
openai
torch